backend/refactor: dashboard driver module decomposition
