+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/dynamic-offer-driver-app.cabal
@@ -136,13 +136,13 @@ library
       Domain.Action.Beckn.Status
       Domain.Action.Beckn.Track
       Domain.Action.Beckn.Update
-      Domain.Action.Dashboard.Driver
-      Domain.Action.Dashboard.Driver.Notification
+      Domain.Action.Dashboard.Common
       Domain.Action.Dashboard.Exotel
       Domain.Action.Dashboard.Fleet.Driver
       Domain.Action.Dashboard.Fleet.Registration
       Domain.Action.Dashboard.Management.Booking
       Domain.Action.Dashboard.Management.Driver
+      Domain.Action.Dashboard.Management.Driver.Notification
       Domain.Action.Dashboard.Management.DriverCoins
       Domain.Action.Dashboard.Management.DriverGoHome
       Domain.Action.Dashboard.Management.DriverReferral
@@ -152,6 +152,7 @@ library
       Domain.Action.Dashboard.Management.NammaTag
       Domain.Action.Dashboard.Management.Revenue
       Domain.Action.Dashboard.Management.Ride
+      Domain.Action.Dashboard.Management.Subscription
       Domain.Action.Dashboard.Overlay
       Domain.Action.Dashboard.Ride
       Domain.Action.Dashboard.RideBooking.Driver
