+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src/Domain/Action/Dashboard/Management/Driver/Notification.hs
@@ -12,7 +12,7 @@
  the GNU Affero General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
 -}
 
-module Domain.Action.Dashboard.Driver.Notification
+module Domain.Action.Dashboard.Management.Driver.Notification
   ( sendDummyRideRequestToDriver,
     triggerDummyRideRequest,
   )
