+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src/Domain/Action/UI/Driver.hs
@@ -97,7 +97,7 @@ import qualified Data.Text as T
 import Data.Time (Day, diffDays, fromGregorian)
 import qualified Data.Tuple.Extra as DTE
 import Domain.Action.Beckn.Search
-import Domain.Action.Dashboard.Driver.Notification as DriverNotify (triggerDummyRideRequest)
+import Domain.Action.Dashboard.Management.Driver.Notification as DriverNotify (triggerDummyRideRequest)
 import qualified Domain.Action.UI.DriverGoHomeRequest as DDGR
 import qualified Domain.Action.UI.DriverHomeLocation as DDHL
 import Domain.Action.UI.DriverOnboarding.AadhaarVerification (fetchAndCacheAadhaarImage)
